[BUGFIX] reduce the amount of abstraction
The current abstraction leeds to issues with rules using more than just the super basic features. The Rule has been rewritten, to still extract some components like the source and destination address, but also provide the original command. Also the API has changed, pass in an iptables rule as a string, rather than a struct.

Still, tables are not supported yet, but I'll be adding it as soon as there is time.